requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
pynmea2==1.15.0
pyserial==3.4
requests==2.22.0
urllib3==1.25.7
